requirements.txt
openai>=1.0.0
requests>=2.28.0
python-dotenv>=0.20.0
rich>=12.0.0
tabulate>=0.8.10
tenacity>=8.0.0
networkx>=2.8.0
numpy>=1.22.0
python-louvain>=0.16 